Midterm final submission (Read description)
I wasn't able to figure out how to use a second API like Google Search or Gemini for a description. I thought it would be easy but it ended up being a nightmare, I had to make new accounts in like 3 different places, had to write hundreds of lines of unfamiliar code, and then got endless error messages that I had no idea how to fix. Local servers, proxies, it was a mess, and it also slowed down the site like crazy. Not worth the trouble at all, unfortunately.